requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.4
itsdangerous==0.24
pika==0.10.0
wsgiref==0.1.2
